+++ MotorRotationEnum.cs
@@ -1,0 +1,9 @@
+﻿
+namespace YellowImpactPitchIndication {
+	public enum MotorRotationEnum {
+		Movement_Speed_Only = 0,
+		Time_Left_Only = 1,
+		Max_of_Speed_and_Time_Left = 2,
+		Both_Speed_and_Time_Left = 3
+	}
+}