+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog 
 
+## v1.0.2
+- Added an option to make the motor's rotation show yellow hit cooldown time 
+- Split pitch, volume, and opacity options for yellow hit cooldown steam by yellow hit count 
+- Added yellow hit cooldown steam particle count options by yellow hit count 
+- Steam volume and particle count now increases with yellow hit count by default 
+- Renamed "Cooldown Particle Color" and "Cooldown Particle Opacity" to "Long Cooldown Particle Color" and "Long Cooldown Particle Opacity"
+- Split "Disable on long cooldown" into "Disable audio on long cooldown" and "Disable particles on long cooldown"
+- Added headers to the config 
+
 ## v1.0.1 
 - Fixed a bug where the steam would not reenable itself when swapping to a hammer that had steam enabled the last time it was held 
 - Added an option to have both the yellow timer steam and long cooldown steam enabled at once